+++ azext_iot/product/test/command_tests.py
@@ -278,9 +278,8 @@ def _read_certificate_from_file(certificate_path):
     with open(file=certificate_path, mode="rb") as f:
         data = f.read()
 
-        from base64 import encodestring  # pylint: disable=no-name-in-module
-
-        return encodestring(data)
+        from base64 import b64encode  # pylint: disable=no-name-in-module
+        return b64encode(data).decode('utf-8')
 
 
 def _create_from_file(configuration_file):